Move the advanced dialogs to a separate file
